backend/requirements.txt
flask
flask-cors
faiss-cpu
sentence-transformers
pypdf
google-generativeai
numpy
python-dotenv
requests